refactor: split DirectionsButton into its own component
