+++ assets/css/_directions_button.scss
@@ -1,0 +1,9 @@
+.c-directions-button {
+  @include button-primary;
+  @include button-icon(1rem);
+  margin-bottom: 1rem;
+  display: inline-flex;
+  align-items: center;
+  gap: 4px;
+  text-decoration: none;
+}
