+++ assets/css/app.scss
@@ -65,6 +65,7 @@ $list-group-border-color: $color-gray-300;
 @import "crowding";
 @import "cutout_overlay";
 @import "data_status_banner";
+@import "directions_button";
 @import "disconnected_modal";
 @import "drawer_tab";
 @import "filter_accordion";
